Work through the type checker's findings without changing behaviour
mypy reported 581 errors after the third-party stub overrides; it now reports
234, with the test suite unchanged at 410 passing.

Annotations that disagreed with the code they describe:
- `Filter.get_coefficients` was typed `list | tuple | NDArray | None` while
  every branch returns or raises, so every caller that unpacked it produced a
  "None is not iterable" error. It is now overloaded on
  `FilterCoefficientsType`, as is `tools.fractional_octave_frequencies` on
  `return_cutoff`, so callers get the type their argument selects.
- The five `save_*` methods were typed `-> None` and returned `self`.
- The private `__set_parameters` of `Compressor`, `Tremolo`, `Chorus` and
  `DigitalDelay` took `float` while their public `set_parameters` passes
  `None` for "leave unchanged".
- `_fractional_octave_smoothing` took an `int` octave fraction from callers
  that all pass floats, `Filter.from_zpk` took real zeros and poles,
  `_reconstruct_framed_signal` rejected the `None` safety threshold its body
  handles, `_window_this_ir` rejected a `ParametrizedWindow`, and
  `Distortion`'s levels rejected the scalars they default to.

`DigitalDelay.set_parameters` raised a `TypeError` unless both parameters were
passed, contradicting its own docstring; it now leaves a `None` unchanged, and
a test covers it.

Everything else is narrowing that the checker could not do on its own:
variables rebound from a parameter to a different type get a second name
(`LRFilterBank`'s coefficients, `_check_ir_start_reverb`, the chirp frequency
range, the plotting helpers' axes, the mode branches of the filter bank
plots), lazily computed attributes are annotated and asserted, `None`-able
values are read once into a local, and the 3D axes are narrowed to `Axes3D`.
`compute_transfer_function` calls `_welch` through a local closure instead of
splatting a dictionary.

+++ dsptoolbox/standard/_standard_backend.py
@@ -327,8 +327,8 @@ def _kaiser_window_fractional(
 def _indices_above_threshold_dbfs(
     time_vec: NDArray[np.float64],
     threshold_dbfs: float,
-    attack_smoothing_coeff: int,
-    release_smoothing_coeff: int,
+    attack_smoothing_coeff: float,
+    release_smoothing_coeff: float,
     normalize: bool = True,
 ) -> NDArray[np.bool_]:
     """Returns indices with power above a given power threshold (in dBFS) in a
@@ -341,9 +341,9 @@ def _indices_above_threshold_dbfs(
         take one channel.
     threshold_dbfs : float
         Threshold in dBFS to be regarded for activation.
-    attack_smoothing_coeff : int
+    attack_smoothing_coeff : float
         Coefficient for attack smoothing for level computation.
-    release_smoothing_coeff : int
+    release_smoothing_coeff : float
         Coefficient for release smoothing for level computation.
     normalize : bool, optional
         When `True`, signal is normalized such that the threshold is relative
